# Speed up DeepCopy