```cs
|    Method |      Mean |     Error |    StdDev | Rank | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------- |----------:|----------:|----------:|-----:|------------:|------------:|------------:|--------------------:|
|  DeepCopy | 183.98 us | 1.2403 us | 1.0995 us |    2 |     10.7422 |           - |           - |            33.37 KB |
| DeepCopy2 |  66.64 us | 0.1954 us | 0.1632 us |    1 |      5.9814 |           - |           - |            18.39 KB |
```

This change speeds up DeepCopy by a factor of 2.76x while reducing memory
allocation by 1.8x

EDIT: after more improvements
```cs
|    Method |      Mean |     Error |    StdDev | Rank | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------- |----------:|----------:|----------:|-----:|------------:|------------:|------------:|--------------------:|
|  DeepCopy | 195.03 us | 3.8461 us | 4.1153 us |    3 |     10.7422 |           - |           - |            33.37 KB |
| DeepCopy2 |  72.02 us | 1.3736 us | 1.2849 us |    2 |      5.9814 |           - |           - |            18.39 KB |
| DeepCopy3 |  39.31 us | 0.7839 us | 0.8388 us |    1 |      4.2725 |           - |           - |             13.2 KB |
```